agent-service/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-huggingface
transformers
torch
accelerate
bitsandbytes
python-dotenv
pydantic-settings
